Add test for GC (only source with negative dec)
